[FIX] Robo should not follow to the bundle
