--- src/ExEss/Bundle/CmsBundle/Robo/Task/ExecAware.php
@@ -1,22 +1,0 @@
-<?php
-
-namespace ExEss\Bundle\CmsBundle\Robo\Task;
-
-use Robo\Collection\CollectionBuilder;
-use Robo\Robo;
-use Robo\Task\Base\loadTasks;
-use Robo\TaskAccessor;
-
-trait ExecAware
-{
-    use TaskAccessor;
-    use loadTasks;
-
-    /**
-     * Scaffold the collection builder
-     */
-    public function collectionBuilder(): CollectionBuilder
-    {
-        return Robo::createDefaultContainer()->get('collectionBuilder', [new \Robo\Tasks]);
-    }
-}
