+++ src/ExEss/Cms/Robo/Task/Db/Release/Import/Translations.php
@@ -1,0 +1,24 @@
+<?php
+
+namespace ExEss\Cms\Robo\Task\Db\Release\Import;
+
+use ExEss\Cms\Robo\Task\Db\SqlFileImport;
+
+class Translations extends AbstractDbImport
+{
+    protected string $subPath = '/translations';
+
+    /**
+     * @inheritdoc
+     */
+    public function runImport()
+    {
+        (
+            new SqlFileImport(
+                $this->output(),
+                \glob(self::RELEASE_DIR . $this->subPath . "/*.sql"),
+                $this->getDatabase()
+            )
+        )->run();
+    }
+}
